Created a working fps loop.
The new fps works like this: There is a thread inside game draw. It
uses the phone's real world time in milliseconds to record the
timestamp of the last draw. Then, whenever that timestamp is smaller
than the current time minus the time for a frame, the views draw method
is called.